Changed rssi and battery parameters type to sbyte;

+++ src/HideezClient/HideezClient/Messages/Remote/Remote_BatteryChangedMessage.cs
@@ -2,12 +2,12 @@
 {
     class Remote_BatteryChangedMessage : Remote_BaseMessage
     {
-        public Remote_BatteryChangedMessage(string serialNo, int battery)
+        public Remote_BatteryChangedMessage(string serialNo, sbyte battery)
             : base(serialNo)
         {
             Battery = battery;
         }
 
-        public int Battery { get; }
+        public sbyte Battery { get; }
     }
 }
