+++ src/HideezService/HideezServiceHost/HideezServiceCallbacks.cs
@@ -38,11 +38,11 @@ namespace HideezServiceHost
         {
         }
 
-        public void RemoteConnection_BatteryChanged(string serialNo, int battery)
+        public void RemoteConnection_BatteryChanged(string serialNo, sbyte battery)
         {
         }
 
-        public void RemoteConnection_RssiReceived(string serialNo, double rssi)
+        public void RemoteConnection_RssiReceived(string serialNo, sbyte rssi)
         {
         }
